Update New-MoodleUser.ps1
Hello! We utilize the Office 365 add in to manage our authentication via OIDC/Open ID Connect. I'd love to use this with your scripts as a possible authentication method. I am a novice coder and not sure if further changes would be needed, beyond just allowing it as an option.